A means to provide custom UA_gethostname
When using lwip, there's often system-specific concerns regarding the
notion of a "hostname" as far as exposing it ot the outside world e.g.
via OPC-UA.  The existing implementation of UA_gethostname for lwip
might work fine in environments where DNS actually works AND where
lwip's netif-specific hostname function is in use, but those can be bad
assumptions.

Allow the integrator to define UA_GETHOSTNAME_PROVIDED and provide their
own linked-in implementation of UA_gethostname.  This appears to have
been the intent of the preprocessor logic in arch/common/ua_lwip.h,
though it wasn't actually possible to do so.